requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
